+++ examples/solidstart/src/server/mutations.ts
@@ -1,4 +1,4 @@
-import { mutation$, redirect } from '@prpc/solid'
+import { mutation$, redirect$ } from '@prpc/solid'
 import { z } from 'zod'
 
 export const add = mutation$(
@@ -7,7 +7,7 @@ export const add = mutation$(
     await new Promise((res) => setTimeout(res, 250))
     const result = payload.a + payload.b
     if (result === 10) {
-      return redirect('/reached-10')
+      return redirect$('/reached-10')
     }
     return result
   },