+++ packages/solid/src/query.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { createQuery, type CreateQueryResult } from '@adeora/solid-query'
+import { useNavigate } from 'solid-start'
 import type { z, ZodObject } from 'zod'
 import type {
   ExpectedFn,
@@ -8,7 +9,7 @@ import type {
   AsParam,
   ModifQueryOptions,
 } from './types'
-import { genQueryKey, getNewOpts, tryAndWrap, unwrapValue } from './utils'
+import { genQueryKey, tryAndWrap, unwrapValue } from './utils'
 
 export function query$<
   ZObj extends ZodObject<any> | undefined,
@@ -19,11 +20,12 @@ export function query$<
     input: AsParam<Fn>,
     queryOpts?: ModifQueryOptions<FCreateQueryOptions<InferReturnType<Fn>>>
   ) => {
-    const newOpts = getNewOpts(queryOpts)
+    const navigate = useNavigate()
     return createQuery(() => ({
       queryKey: genQueryKey(key, unwrapValue(input)),
-      queryFn: () => tryAndWrap(queryFn, input, newOpts?.().alwaysCSRRedirect),
-      ...((newOpts?.() || {}) as any),
+      queryFn: () =>
+        tryAndWrap(queryFn, input, navigate, queryOpts?.().alwaysCSRRedirect),
+      ...((queryOpts?.() || {}) as any),
     })) as CreateQueryResult<InferReturnType<Fn>>
   }
 }